Add files via upload
This study is based to inspect the ability to predict housing prices is important such that investors can make better asset allocation decisions, including the pricing and underwriting of the mortgage. The study intends to explore the possibility of modelling housing price based on Asset-based. Asset pricing theory all stems from one simple concept: price equals expected discounted payoff (rent). The model assumes an association among user cost UCt, the house selling prices, Pt, and the rents Rt
In this project, first, the data is collected online (Google Doc) via a self-administered questionnaire through the real estate investors in Delhi/NCR and have stored dataset in the database of MS-Access. The dataset has fifteen variables, with twelve analysis variables, which were of both types: qualitative and quantitative. The study was intended to find the relationship of the independent variable (Indicative price) with the dependent variables such as Payoff (Rent), Annual Maintenance Cost and the expected rental increase.
The data was analyzed using CRISP approach. The result was analyzed statistically, to come up to the conclusion that Asset-Based Models can be applied in the case of NCR properties.
The Algorithm has been built in R-Script, and the Model is compared between the two, Simple Multiple Linear Regression and Stepwise Multiple Linear Regression and the results were compared.